Rename UpgradeButton component to SubscriptionButton and update imports accordingly
