+++ frontend/src/components/quota/quota-usage-display.tsx
@@ -5,7 +5,7 @@ import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Progress } from "@/components/ui/progress";
 import { Badge } from "@/components/ui/badge";
 import { Separator } from "@/components/ui/separator";
-import { UpgradeButton } from "@/components/layout/upgrade-button";
+import { UpgradeButton } from "@/components/layout/subscription-button";
 
 interface QuotaUsageDisplayProps {
   quotaUsage: QuotaUsage;
